Adds stringloader for word list and update card
Adds the ability for a .txt file to be read from a url, generating a
string that is then split into a list. Currently that list is stored
on the card object itself. It checks to see if the word list is already
populated, and also calls a function to pick a random word from that
list.